app/requirements.txt
Jinja2 >= 2.7.3
PyYAML >= 3.11
nose >= 1.3.7
